I've rewritten almost completely the distribution simulation and extracted prediction simulation as a separate component. The second one now uses demand and supply prediction functions that are computed as cumulative distribution of data colleced from other flow actions. I've also introduced separate way of predicting the supply for actions that have already begun and those that that are yet to start. I have to accomodate the redistribution simulation to the new structure.

+++ Model/CityBikes.Flow/relocation/relocation-unit.hpp
@@ -14,7 +14,7 @@ namespace CityBikes::Flow::Relocation
 		size_t destination;
 
 		/// <summary> When is destination going to be reached </summary>
-		size_t timeUntilDestination;
+		float timeUntilDestination;
 
 		RelocationUnit(size_t position, size_t currentLoad, size_t maxLoad) :
 			destination(position),
